[fix] memberservice to authservice
