+++ src/main/java/com/academy/sivillageclonebe/admin/controlloer/CategoryController.java
@@ -1,15 +1,12 @@
-package com.academy.sivillageclonebe.category.controlloer;
+package com.academy.sivillageclonebe.admin.controlloer;
 
-import com.academy.sivillageclonebe.category.dto.MainCategoryRequestDto;
-import com.academy.sivillageclonebe.category.service.CategoryService;
-import com.academy.sivillageclonebe.common.entity.CommonResponseEntity;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/api/v1/category")
+@RequestMapping("/api/v1/admin/category")
 public class CategoryController {
 
 //    public final CategoryService categoryService;
