+++ src/main/java/com/academy/sivillageclonebe/admin/reponstity/DetailedCategoriesRepository.java
@@ -1,0 +1,4 @@
+package com.academy.sivillageclonebe.admin.reponstity;
+
+public interface DetailedCategoriesRepository {
+}
