+++ src/main/java/com/academy/sivillageclonebe/admin/reponstity/MainCategoriesRepository.java
@@ -1,0 +1,4 @@
+package com.academy.sivillageclonebe.admin.reponstity;
+
+public interface MainCategoriesRepository {
+}
