+++ src/main/java/com/academy/sivillageclonebe/admin/service/CategoryService.java
@@ -1,0 +1,8 @@
+package com.academy.sivillageclonebe.admin.service;
+
+import com.academy.sivillageclonebe.admin.dto.MainCategoryRequestDto;
+
+public interface CategoryService {
+
+    void createMainCategories(MainCategoryRequestDto mainCategoryRequestDto);
+}
