+++ src/main/java/com/academy/sivillageclonebe/admin/service/CategoryServiceImpl.java
@@ -1,0 +1,14 @@
+package com.academy.sivillageclonebe.admin.service;
+
+import com.academy.sivillageclonebe.admin.dto.MainCategoryRequestDto;
+
+public class CategoryServiceImpl implements CategoryService{
+
+
+    @Override
+    public void createMainCategories(MainCategoryRequestDto mainCategoryRequestDto) {
+
+
+
+    }
+}
