+++ src/main/java/com/academy/sivillageclonebe/admin/vo/MainCategoryResponseVo.java
@@ -1,0 +1,4 @@
+package com.academy.sivillageclonebe.admin.vo;
+
+public class MainCategoryResponseVo {
+}
