--- src/main/java/com/academy/sivillageclonebe/category/reponstity/MainCategoriesRepository.java
@@ -1,4 +1,0 @@
-package com.academy.sivillageclonebe.category.reponstity;
-
-public interface MainCategoriesRepository {
-}
