--- src/main/java/com/academy/sivillageclonebe/category/service/CategoryService.java
@@ -1,9 +1,0 @@
-package com.academy.sivillageclonebe.category.service;
-
-import com.academy.sivillageclonebe.category.dto.MainCategoryRequestDto;
-import com.academy.sivillageclonebe.category.vo.MainCategoryRequestVo;
-
-public interface CategoryService {
-
-    void createMainCategories(MainCategoryRequestDto mainCategoryRequestDto);
-}
