--- src/main/java/com/academy/sivillageclonebe/category/service/CategoryServiceImpl.java
@@ -1,15 +1,0 @@
-package com.academy.sivillageclonebe.category.service;
-
-import com.academy.sivillageclonebe.category.dto.MainCategoryRequestDto;
-import com.academy.sivillageclonebe.category.reponstity.MainCategoriesRepository;
-
-public class CategoryServiceImpl implements CategoryService{
-
-
-    @Override
-    public void createMainCategories(MainCategoryRequestDto mainCategoryRequestDto) {
-
-
-
-    }
-}
