--- src/main/java/com/academy/sivillageclonebe/category/vo/MainCategoryResponseVo.java
@@ -1,4 +1,0 @@
-package com.academy.sivillageclonebe.category.vo;
-
-public class MainCategoryResponseVo {
-}
