+++ src/main/java/com/academy/sivillageclonebe/member/controller/AuthController.java
@@ -3,7 +3,9 @@ package com.academy.sivillageclonebe.member.controller;
 import com.academy.sivillageclonebe.common.entity.CommonResponseEntity;
 import com.academy.sivillageclonebe.common.entity.CommonResponseMessage;
 import com.academy.sivillageclonebe.member.dto.SignInRequestDto;
+import com.academy.sivillageclonebe.member.dto.SignInResponseDto;
 import com.academy.sivillageclonebe.member.dto.SignUpRequestDto;
+import com.academy.sivillageclonebe.member.service.AuthService;
 import com.academy.sivillageclonebe.member.service.MemberService;
 import com.academy.sivillageclonebe.member.vo.SignInRequestVo;
 import com.academy.sivillageclonebe.member.vo.SignInResponseVo;
@@ -18,26 +20,22 @@ import org.springframework.web.bind.annotation.*;
 @Slf4j
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/user")
+@RequestMapping("/api/v1/auth")
 public class AuthController {
 
-    private final MemberService memberService;
-    private final ModelMapper modelMapper;
+    private final AuthService authService;
 
     @Operation(summary = "SignIn API", description = "SignIn API 입니다.", tags = {"Auth"})
     @PostMapping("/sign-in")
     public CommonResponseEntity<SignInResponseVo> signIn(
             @RequestBody SignInRequestVo signInRequestVo) {
-        SignInRequestDto signInRequestDto = SignInRequestDto.builder().
-                email(signInRequestVo.getEmail()).
-                password(signInRequestVo.getPassword()).
-                build();
-        SignInResponseVo signInResponseVo = modelMapper.map(memberService.signIn(signInRequestDto), SignInResponseVo.class);
-        log.info("signInResponseVo : {}", signInResponseVo);
+
         return new CommonResponseEntity<>(
                 HttpStatus.OK,
                 CommonResponseMessage.SUCCESS.getMessage(),
-                signInResponseVo);
+                authService.signIn(signInRequestVo.toDto()).toVo()
+        );
+
 
     }
 
@@ -45,9 +43,13 @@ public class AuthController {
     @PostMapping("/sign-up")
     public CommonResponseEntity<Void> signUp(
             @RequestBody SignUpRequestVo signUpRequestVo) {
-        log.info("signUpRequestVo : {}", signUpRequestVo);
-        SignUpRequestDto signUpRequestDto = modelMapper.map(signUpRequestVo, SignUpRequestDto.class);
-        memberService.signUp(signUpRequestDto);
-        return new CommonResponseEntity<>(HttpStatus.OK, CommonResponseMessage.SUCCESS.getMessage(), null);
+
+        authService.signUp(signUpRequestVo.toDto());
+
+        return new CommonResponseEntity<>(
+                HttpStatus.OK,
+                CommonResponseMessage.SUCCESS.getMessage(),
+                null
+        );
     }
 }
