+++ src/main/java/com/academy/sivillageclonebe/member/dto/SignInRequestDto.java
@@ -1,5 +1,6 @@
 package com.academy.sivillageclonebe.member.dto;
 
+import com.academy.sivillageclonebe.member.vo.SignInRequestVo;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
