+++ src/main/java/com/academy/sivillageclonebe/member/vo/SignInRequestVo.java
@@ -1,5 +1,7 @@
 package com.academy.sivillageclonebe.member.vo;
 
+import com.academy.sivillageclonebe.member.dto.SignInRequestDto;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
 
@@ -9,5 +11,13 @@ public class SignInRequestVo {
 
     private String email;
     private String password;
+
+    @Builder
+    public SignInRequestDto toDto() {
+        return SignInRequestDto.builder()
+                .email(email)
+                .password(password)
+                .build();
+    }
 
 }
