+++ src/main/java/com/academy/sivillageclonebe/member/vo/SignUpRequestVo.java
@@ -1,7 +1,9 @@
 package com.academy.sivillageclonebe.member.vo;
 
+import com.academy.sivillageclonebe.member.dto.SignUpRequestDto;
 import com.academy.sivillageclonebe.member.entity.Oauth;
 import com.academy.sivillageclonebe.member.entity.Role;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
 
@@ -16,5 +18,16 @@ public class SignUpRequestVo {
     private String password;
     private String name;
     private String phone;
+
+    @Builder
+    public SignUpRequestDto toDto(){
+        return SignUpRequestDto.builder()
+                .role(role)
+                .email(email)
+                .password(password)
+                .name(name)
+                .phone(phone)
+                .build();
+    }
 
 }
